refactor(editor): colocate TipTapEditorProps with its component
Move the TipTapEditorProps prop-type out of the shared types.ts and directly
above TipTapEditor in components.tsx, its only consumer. types.ts keeps the
genuinely shared types (highlight targets, CommentDraft, CoeditorHandle) and
sheds the react/tiptap/yjs/awareness imports that existed only to type the
props.
